requirements.txt
Django~=5.2
django-filter~=25.1
django-markdown~=0.8
django-prose-editor~=0.10
djangorestframework~=3.16
django-vite~=3.1
django-cors-headers
python-dotenv